+++ packages/adapter-interfaces/src/adapters/Keys.ts
@@ -1,8 +1,13 @@
-import { SigningKey } from "../types";
+import { SigningKey, Totals } from "../types";
+import { ListParams } from "../types/ListParams";
+
+export interface ListKeysResponse extends Totals {
+  signingKeys: SigningKey[];
+}
 
 export interface KeysAdapter {
   create: (key: SigningKey) => Promise<void>;
-  list: () => Promise<SigningKey[]>;
+  list: (params?: ListParams) => Promise<ListKeysResponse>;
   update: (
     kid: string,
     key: Partial<Omit<SigningKey, "kid">>,