+++ packages/kysely/migrate/migrations/2023-12-26T10:58:00_signing-keys.ts
@@ -8,17 +8,21 @@ export async function up(db: Kysely<Database>): Promise<void> {
     .addColumn("tenant_id", "varchar(255)", (col) =>
       col.references("tenants.id").onDelete("cascade"),
     )
-    // This will be removed in a future migration
-    // .addColumn("private_key", "varchar(8192)", (col) => col.notNull())
-    // .addColumn("public_key", "varchar(1024)", (col) => col.notNull())
+
     .addColumn("created_at", "varchar(255)", (col) => col.notNull())
     .addColumn("revoked_at", "varchar(255)")
-    .addColumn("cert", "varchar(2048)")
-    .addColumn("pkcs7", "varchar(2048)")
+    .addColumn("cert", "varchar(4096)")
+    .addColumn("pkcs7", "varchar(4096)")
     .addColumn("fingerprint", "varchar(256)")
     .addColumn("thumbprint", "varchar(256)")
     .addColumn("current_since", "varchar(256)")
     .addColumn("current_until", "varchar(256)")
+    .addColumn("type", "varchar(50)", (col) =>
+      col.notNull().defaultTo("jwt_signing"),
+    )
+    .addColumn("connection", "varchar(255)", (col) =>
+      col.references("connections.id").onDelete("cascade"),
+    )
     .execute();
 }