+++ packages/kysely/src/tenants/list.ts
@@ -5,33 +5,29 @@ import { ListParams } from "@authhero/adapter-interfaces";
 import getCountAsInt from "../utils/getCountAsInt";
 
 export function list(db: Kysely<Database>) {
-  return async (
-    params: ListParams = {
-      page: 0,
-      per_page: 50,
-      include_totals: false,
-    },
-  ) => {
+  return async (params: ListParams) => {
     let query = db.selectFrom("tenants");
 
-    if (params.sort && params.sort.sort_by) {
+    const { page = 0, per_page = 50, include_totals = false, sort, q } = params;
+
+    if (sort && sort.sort_by) {
       const { ref } = db.dynamic;
-      query = query.orderBy(ref(params.sort.sort_by), params.sort.sort_order);
+      query = query.orderBy(ref(sort.sort_by), sort.sort_order);
     }
 
-    if (params.q) {
-      query = query.where((eb) => eb.or([eb("name", "like", `%${params.q}%`)]));
+    if (q) {
+      query = query.where((eb) => eb.or([eb("name", "like", `%${q}%`)]));
     }
 
-    const filteredQuery = query
-      .offset(params.page * params.per_page)
-      .limit(params.per_page);
+    const filteredQuery = query.offset(page * per_page).limit(per_page);
 
     const tenants = await filteredQuery.selectAll().execute();
 
-    if (!params.include_totals) {
+    const mappedTenants = tenants.map(removeNullProperties);
+
+    if (!include_totals) {
       return {
-        tenants,
+        tenants: mappedTenants,
       };
     }
 
@@ -42,9 +38,9 @@ export function list(db: Kysely<Database>) {
     const countInt = getCountAsInt(count);
 
     return {
-      tenants: tenants.map(removeNullProperties),
-      start: (params.page - 1) * params.per_page,
-      limit: params.per_page,
+      tenants: mappedTenants,
+      start: page * per_page,
+      limit: per_page,
       length: countInt,
     };
   };